# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: CodeQuest API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /submissions:
    post:
      operationId: submitCode
      tags: [submissions]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitCodeRequest"
      responses:
        "200":
          description: Submission result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /submissions/run:
    post:
      operationId: runCode
      tags: [submissions]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunCodeRequest"
      responses:
        "200":
          description: Run result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /submissions/history:
    get:
      operationId: getSubmissionHistory
      tags: [submissions]
      security:
        - bearerAuth: []
      parameters:
        - name: problemId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Submission history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionHistoryResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    Badge:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        category:
          type: string
          enum: [solving, xp, streak, quest, level]
        rarity:
          type: string
          enum: [common, rare, epic, legendary]
        earned:
          type: boolean
        earnedAt:
          type: string
          format: date-time
          nullable: true
      required: [slug, name, description, icon, category, rarity, earned]
...
    SubmitCodeRequest:
      type: object
      properties:
        problemId:
          type: integer
        code:
          type: string
        language:
          type: string
          enum: [python, javascript]
      required: [problemId, code, language]
    RunCodeRequest:
      type: object
      properties:
        problemId:
          type: integer
        code:
          type: string
        language:
          type: string
          enum: [python, javascript]
      required: [problemId, code, language]
    TestCaseResult:
      type: object
      properties:
        passed:
          type: boolean
        input:
          type: string
        expectedOutput:
          type: string
        actualOutput:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        time:
          type: number
          nullable: true
      required: [passed, input, expectedOutput]
    SubmissionResult:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [Accepted, Wrong Answer, Time Limit Exceeded, Runtime Error, Compilation Error]
        passedCount:
          type: integer
        totalCount:
          type: integer
        xpEarned:
          type: integer
        bonusXpEarned:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/TestCaseResult"
        runtime:
          type: number
          nullable: true
        newBadges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"
        streakUpdated:
          type: boolean
        newStreak:
          type: integer
        newStarRank:
          type: integer
          nullable: true
        isDailyQuest:
          type: boolean
        levelUp:
          type: boolean
        newLevel:
          type: integer
          nullable: true
      required:
        - id
        - status
        - passedCount
        - totalCount
        - xpEarned
        - bonusXpEarned
        - results
        - newBadges
        - streakUpdated
        - newStreak
        - isDailyQuest
        - levelUp
    RunResult:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/TestCaseResult"
        allPassed:
          type: boolean
      required: [results, allPassed]
    Submission:
      type: object
      properties:
        id:
          type: integer
        problemId:
          type: integer
        problemTitle:
          type: string
        language:
          type: string
        status:
          type: string
        passedCount:
          type: integer
        totalCount:
          type: integer
        runtime:
          type: number
          nullable: true
        submittedAt:
          type: string
          format: date-time
      required: [id, problemId, problemTitle, language, status, passedCount, totalCount, submittedAt]
    SubmissionHistoryResponse:
      type: object
      properties:
        submissions:
          type: array
          items:
            $ref: "#/components/schemas/Submission"
      required: [submissions]
...