# lib/api-zod/src/generated/types/testCaseResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeQuest API specification
 * OpenAPI spec version: 0.2.0
 */

export interface TestCaseResult {
  passed: boolean;
  input: string;
  expectedOutput: string;
  actualOutput?: string | null;
  error?: string | null;
  time?: number | null;
}